requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic-settings==2.13.1

sqlalchemy==2.0.49
alembic==1.18.4

asyncpg==0.31.0

pytest==9.0.2
httpx==0.28.1

black==26.3.1

shortuuid==1.0.13